feat: aql2
