+++ README.md
@@ -26,12 +26,10 @@ This module implements two key interfaces:
 
 Both interfaces can be used independently or together depending on your application's needs. The interfaces are installed separately to maintain modularity and minimize dependencies.
 
-
-| Name          | Install command                         |                                                                   |
-| ------------- | --------------------------------------- | ----------------------------------------------------------------- |
-| MongoDB       | `ajs module imports add mongodb`        | [Documentation](https://github.com/AntelopeJS/interface-mongodb)  |
-| Database      | `ajs module imports add database`       | [Documentation](https://github.com/AntelopeJS/interface-database) |
-
+| Name     | Install command                   |                                                                   |
+| -------- | --------------------------------- | ----------------------------------------------------------------- |
+| MongoDB  | `ajs module imports add mongodb`  | [Documentation](https://github.com/AntelopeJS/interface-mongodb)  |
+| Database | `ajs module imports add database` | [Documentation](https://github.com/AntelopeJS/interface-database) |
 
 ## Overview
