+++ output/database/beta/common.d.ts
