+++ output/database/beta/query.d.ts
