+++ output/database/beta/stream.d.ts
