+++ output/database/beta/valueproxy.d.ts
