--- output/database/beta2/index.d.ts
@@ -1,6 +1,0 @@
-export { ValueProxy, ValueProxyOrValue } from './valueproxy';
-export { Query } from './query';
-export { Datum } from './datum';
-export { Stream } from './stream';
-export { Selection, SingleSelection, Table } from './selection';
-export { Schema, SchemaInstance, SchemaDefinition } from './schema';
