+++ output/mongodb/beta/index.d.ts
@@ -7,38 +7,6 @@
  */
 import { MongoClient } from 'mongodb';
 /**
- * Internal namespace for MongoDB client management
- *
- * Contains implementation details that should not be directly accessed by application code.
- * This namespace handles the MongoDB client connection lifecycle and state.
- */
-export declare namespace internal {
-    /**
-     * Promise that resolves to the MongoDB client instance once connected
-     * Acts as a singleton access point to the database connection
-     */
-    let client: Promise<MongoClient>;
-    /**
-     * Connection state flag
-     * Indicates whether the MongoDB client is currently connected
-     */
-    let connected: boolean;
-    /**
-     * Client setter function
-     * Used internally to resolve the client promise when a connection is established
-     *
-     * @param client The MongoDB client instance to set
-     */
-    let SetClient: (client: MongoClient) => void;
-    /**
-     * Resets the client promise and prepares for a new connection
-     * Used when disconnecting or when needing to establish a fresh connection
-     *
-     * @returns A new unresolved client promise
-     */
-    const UnsetClient: () => Promise<MongoClient>;
-}
-/**
  * Returns a promise that resolves to the MongoDB client
  *
  * This is the primary way for application code to access the MongoDB connection.
