--- playground/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "playground",
-  "version": "0.0.1",
-  "main": "dist/index.js",
-  "scripts": {
-    "dev": "ajs project run -w",
-    "build": "tsc"
-  },
-  "devDependencies": {
-    "@types/node": "^20.5.7",
-    "prettier": "^2.6.1",
-    "typescript": "^5.8.3"
-  },
-  "antelopeJs": {
-    "imports": [
-      {
-        "name": "core@beta",
-        "isLocal": true
-      },
-      {
-        "name": "database@beta",
-        "isLocal": true
-      },
-      "logging@beta"
-    ],
-    "baseUrl": "dist",
-    "paths": {
-      "@/*": [
-        "*"
-      ]
-    },
-    "importsOptional": []
-  }
-}
