--- playground/tsconfig.json
@@ -1,26 +1,0 @@
-{
-  "compilerOptions": {
-    "moduleResolution": "node",
-    "target": "esnext",
-    "module": "commonjs",
-    "lib": ["ES2020"],
-    "allowJs": true,
-    "outDir": "dist",
-    "strict": true,
-    "noImplicitAny": true,
-    "esModuleInterop": true,
-    "resolveJsonModule": true,
-    "removeComments": true,
-    "preserveConstEnums": true,
-    "experimentalDecorators": true,
-    "emitDecoratorMetadata": true,
-    "declaration": true,
-    "sourceMap": true,
-    "useDefineForClassFields": true,
-    "baseUrl": "src",
-    "paths": {
-      "@ajs/*": ["../../output/*", "../../.antelope/interfaces.d/*", "../.antelope/interfaces.d/*"]
-    }
-  },
-  "include": ["src/**/*.ts"]
-}
