--- src/implementations/database/beta2/index.ts
@@ -1,25 +1,0 @@
-/*
-
-FEATURES
-
-key/default/map/pluck/without
-filter
-join/lookup
-group
-orderBy
-slice/nth
-count/sum/...
-changes
-
-
-USEFUL
-$group https://www.mongodb.com/docs/manual/reference/operator/aggregation/group/
- -> _id: null for whole stream
-$addToSet https://www.mongodb.com/docs/manual/reference/operator/aggregation/addToSet/
-$push https://www.mongodb.com/docs/manual/reference/operator/aggregation/push/
-$replaceRoot https://www.mongodb.com/docs/manual/reference/operator/aggregation/replaceRoot/
-
-$lookup
-
-
- */
