+++ src/interfaces/database/beta/common.ts
