+++ src/interfaces/database/beta/schema.ts
