+++ src/interfaces/database/beta/stream.ts
