+++ src/interfaces/database/beta/valueproxy.ts
